add support for no_std
